+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] / 2021-12-11
+- Add Scale ImageSource
+- Add RibbonItemExtension
+- Fix net47 to net46
+
 ## [0.0.1] / 2021-12-11
 - Fix Release / Content Build folder
 - Add README / LICENSE
@@ -17,4 +22,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: https://github.com/ricaun-io/ricaun.Revit.UI/compare/1.0.0...HEAD
+[0.0.2]: https://github.com/ricaun-io/ricaun.Revit.UI/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/ricaun-io/ricaun.Revit.UI/compare/0.0.1